Break up FetchService into multiple files
